2.5-3.1x speedup on map generation, fixed memory leaks of plates arra…

+++ test/test_randomness.cpp
@@ -29,9 +29,9 @@ TEST(Randomness, SequenceDoubles)
   long seed = 3;
   SimpleRandom randsource(seed);
 
-  EXPECT_FLOAT_EQ(5.1118433e-05, randsource.next_double());
-  EXPECT_FLOAT_EQ(0.53070194, randsource.next_double());
-  EXPECT_FLOAT_EQ(0.053402752, randsource.next_double());
+  EXPECT_FLOAT_EQ((float)5.1118433e-05, (float)randsource.next_double());
+  EXPECT_FLOAT_EQ(0.53070194f, (float)randsource.next_double());
+  EXPECT_FLOAT_EQ(0.053402752f, (float)randsource.next_double());
 }
 
 TEST(Randomness, Maximum)
@@ -39,5 +39,5 @@ TEST(Randomness, Maximum)
   long seed = 3;
   SimpleRandom randsource(seed);
 
-  EXPECT_FLOAT_EQ(4.2949673e+09, randsource.maximum());
+  EXPECT_FLOAT_EQ((float)4.2949673e+09, (float)randsource.maximum());
 }
